Move Lyrics & LyricsMOdal to TS
